+++ src/components/Card/Card.tsx
@@ -1,136 +1,144 @@
 import * as React from "react";
-import bem from "../../utils/bem";
+import { Box, useMultiStyleConfig, useStyleConfig } from "@chakra-ui/react";
 
 import { CardLayouts } from "./CardTypes";
 import Heading from "../Heading/Heading";
-import Image from "../Image/Image";
+import Image, { ImageProps } from "../Image/Image";
 import { ImageRatios, ImageSizes } from "../Image/ImageTypes";
+import generateUUID from "../../helpers/generateUUID";
 
-interface CardContentProps {
-  /** Optional className that would be applied to the `card-content` element */
-  className?: string;
+interface CardBaseProps {
+  /** Optional value to control the alignment of the text and elements. */
+  center?: boolean;
+  /** Optional value to render the layout in a row or column (default). */
+  layout?: CardLayouts;
 }
 
-interface CardActionsProps {
-  /** Optional boolean value to control visibility of border on the bottom edge of the card actions element */
+interface CardActionsProps extends CardBaseProps {
+  /** Optional boolean value to control visibility of border on the bottom edge
+   * of the card actions element */
   bottomBorder?: boolean;
-  /** Optional className that would be applied to the `card-actions` element */
-  className?: string;
-  /** Modifiers array for use with BEM. See how to work with modifiers and BEM here: http://getbem.com/introduction/ */
-  modifiers?: string[];
-  /** Optional boolean value to control visibility of border on the top edge of the card actions element */
+  /** Optional boolean value to control visibility of border on the top edge of
+   * the card actions element */
   topBorder?: boolean;
 }
 
-export interface CardProps {
-  /** Optional hex color value used to set the card background color */
-  backgroundColor?: string;
-  /** BlockName for use with BEM. See how to work with blockNames and BEM here: http://getbem.com/introduction/ */
-  blockName?: string;
-  /** Optional boolean value to control the visibility of a border around the card */
-  border?: boolean;
-  /** Optional boolean value to control the alignment of the text and elements within the card */
-  center?: boolean;
-  /** ClassName that appears in addition to "card" */
-  className?: string;
-  /** Optional hex color value used to override the default text color */
-  foregroundColor?: string;
-  /** ID that other components can cross reference for accessibility purposes */
-  id?: string;
-  /** Text description of the image; to follow best practices for accessibility, this prop should not be left blank if `imageSrc` is passed */
-  imageAlt?: string;
-  /** Optional value to control the aspect ratio of the card image; default value is `square` */
-  imageAspectRatio?: ImageRatios;
-  /** Optional boolean value to control the position of the card image */
+interface CardImageProps extends CardBaseProps, ImageProps {
+  /** Optional boolean value to control the position of the `CardImage`. */
   imageAtEnd?: boolean;
-  /** Custom image component used in place of DS `Image` component */
-  imageComponent?: JSX.Element;
-  /** Optional value to control the size of the card image */
-  imageSize?: ImageSizes;
-  /** The path to the image displayed with the card */
-  imageSrc?: string;
-  /** Optional value to control the position of the image placeholder; default value is `column` */
-  layout?: CardLayouts;
-  /** Modifiers array for use with BEM. See how to work with modifiers and BEM here: http://getbem.com/introduction/ */
-  modifiers?: string[];
-  /** Optional padding value.  This value should be entered with the same formatting as a CSS padding attribute (ex. `5%`, `20px`, `2rem`).  If omitted, the Card will use the default padding. */
-  padding?: string;
 }
 
-// Not being used in any examples but available for use.
-export const CardImage = Image;
+export interface CardProps extends CardBaseProps {
+  /** Optional hex color value used to set the card background color. */
+  backgroundColor?: string;
+  /** Optional boolean value to control the visibility of a border around
+   * the card. */
+  border?: boolean;
+  /** Optional CSS class name to add. */
+  className?: string;
+  /** Optional hex color value used to override the default text color. */
+  foregroundColor?: string;
+  /** ID that other components can cross reference for accessibility purposes. */
+  id?: string;
+  /** Text description of the image; to follow best practices for accessibility,
+   * this prop should not be left blank if `imageSrc` is passed. */
+  imageAlt?: string;
+  /** Optional value to control the aspect ratio of the `CardIage`; default
+   * value is `square`. */
+  imageAspectRatio?: ImageRatios;
+  /** Optional boolean value to control the position of the `CardImage`. */
+  imageAtEnd?: boolean;
+  /** Custom image component used in place of DS `Image` component. */
+  imageComponent?: JSX.Element;
+  /** Optional value to control the size of the `CardImage`. */
+  imageSize?: ImageSizes;
+  /** The path to the image displayed within the `Card` component. */
+  imageSrc?: string;
+}
+
+/**
+ * The CardImage component is used internally in the `Card` component. It
+ * renders an `Image` component but with overriding styles specific to the
+ * `Card` component.
+ */
+export function CardImage(props: React.ComponentProps<"img"> & CardImageProps) {
+  const {
+    alt,
+    center,
+    component,
+    imageSize,
+    imageAspectRatio,
+    src,
+    imageAtEnd,
+    layout,
+  } = props;
+  // Additional styles to add to the `Image` component.
+  const styles = useStyleConfig("CardImage", {
+    center,
+    imageAtEnd,
+    imageSize,
+    layout,
+  });
+  return (
+    <Image
+      alt={alt}
+      component={component}
+      imageAspectRatio={imageAspectRatio}
+      imageSize={imageSize}
+      src={src}
+      additionalStyles={styles}
+    />
+  );
+}
 
 // CardHeading child-component
-export const CardHeading = (props) => <Heading {...props} />;
+export const CardHeading = Heading;
 
 // CardContent child-component
-export function CardContent(props: React.PropsWithChildren<CardContentProps>) {
-  const { children, className } = props;
-  return (
-    children && (
-      <div className={bem("card-content", [], "", [className])}>{children}</div>
-    )
-  );
+export function CardContent(props: React.PropsWithChildren<{}>) {
+  const { children } = props;
+  const styles = useStyleConfig("CardContent");
+  return children && <Box __css={styles}>{children}</Box>;
 }
 
 // CardActions child-component
 export function CardActions(props: React.PropsWithChildren<CardActionsProps>) {
-  const {
+  const { bottomBorder, children, topBorder, center, layout } = props;
+  const styles = useStyleConfig("CardActions", {
     bottomBorder,
-    children,
-    className,
-    modifiers = [],
     topBorder,
-  } = props;
+    center,
+    layout,
+  });
 
-  bottomBorder && modifiers.push("bottom-border");
-  topBorder && modifiers.push("top-border");
-
-  return (
-    children && (
-      <div className={bem("card-actions", modifiers, "", [className])}>
-        {children}
-      </div>
-    )
-  );
+  return children && <Box __css={styles}>{children}</Box>;
 }
 
 export default function Card(props: React.PropsWithChildren<CardProps>) {
   const {
     backgroundColor,
-    blockName,
-    center,
+    border,
+    center = false,
     children,
     className,
     foregroundColor,
-    id,
+    id = generateUUID(),
+    imageAlt = "",
+    imageAspectRatio = ImageRatios.Square,
     imageAtEnd,
-    layout,
-    border,
-    padding,
-    modifiers = [],
-    imageAlt,
     imageComponent,
-    imageAspectRatio = ImageRatios.Original,
     imageSize = ImageSizes.Default,
     imageSrc,
+    layout = CardLayouts.Column,
   } = props;
-  const baseClass = "card";
   const hasImage = imageSrc || imageComponent;
-  const styles = {};
+  const customColors = {};
   const cardContents = [];
 
-  layout && modifiers.push(layout);
-  border && modifiers.push("with-border");
-  center && modifiers.push("center");
-  hasImage && modifiers.push("has-image");
-  imageAtEnd && modifiers.push("at-end");
+  backgroundColor && (customColors["backgroundColor"] = backgroundColor);
+  foregroundColor && (customColors["color"] = foregroundColor);
 
-  padding && (styles["padding"] = padding);
-  backgroundColor && (styles["backgroundColor"] = backgroundColor);
-  foregroundColor && (styles["color"] = foregroundColor);
-
-  React.Children.map(children, (child: React.ReactElement) => {
+  React.Children.map(children, (child: React.ReactElement, key) => {
     if (
       child.type === CardHeading ||
       child.props.mdxType === "CardHeading" ||
@@ -139,27 +147,42 @@ export default function Card(props: React.PropsWithChildren<CardProps>) {
       child.type === CardActions ||
       child.props.mdxType === "CardActions"
     ) {
-      cardContents.push(child);
+      const elem = React.cloneElement(child, { key, center, layout });
+      cardContents.push(elem);
     }
   });
 
+  const styles = useMultiStyleConfig("Card", {
+    border,
+    center,
+    hasImage,
+    imageAtEnd,
+    layout,
+  });
+
   return (
-    <div
-      className={bem(baseClass, modifiers, blockName, [className])}
+    <Box
+      className={className}
       id={id}
-      style={styles}
+      __css={{
+        ...styles,
+        ...customColors,
+      }}
     >
       {hasImage && (
         <CardImage
-          className={`card__image ${imageSize}`}
           src={imageSrc ? imageSrc : null}
           component={imageComponent}
-          alt={imageAlt ? imageAlt : null}
+          alt={imageAlt}
           imageSize={imageSize}
           imageAspectRatio={imageAspectRatio}
+          imageAtEnd={imageAtEnd}
+          layout={layout}
         />
       )}
-      <div className={bem("body", [], baseClass)}>{cardContents}</div>
-    </div>
+      <Box className="card-body" __css={styles.body}>
+        {cardContents}
+      </Box>
+    </Box>
   );
 }