+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Composes the DS `HelperErrorText` component with Chakra.
 - Adds `useNYPLTheme` hook for consuming applications to use NYPL DS values.
 - Composes the DS `Image` component with Chakra styling and all `Image` references in other DS components.
+- Composes the DS `Card` component with Chakra.
 
 ## 0.25.1 (October 14, 2021)